alexa/Story_Seeker/lambda/requirements.txt
boto3==1.9.216
ask-sdk-core==1.11.0
psycopg2-binary==2.8.4
python-dotenv==0.10.3
